Remove: 테스트 코드를 패키지 내부로 이동
